auth_helper.py
import streamlit_authenticator as stauth

def setup_auth(config):
    return stauth.Authenticate(
        credentials=config['credentials'],
        cookie_name=config['cookie']['name'],
        key=config['cookie']['key'],
        cookie_expiry_days=config['cookie']['expiry_days']
    )
