requirements.txt
streamlit==1.45.1
PyYAML==6.0.1	
bcrypt==4.3.0
streamlit-option-menu==0.3.13
pandas==2.2.2
plotly==5.22.0
google-api-python-client==2.170.0
google-auth==2.40.2
requests==2.32.3
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-core==2.24.2